mcp-servers/mcp-geo-destinations/requirements.txt
amadeus>=11.0.0
python-dotenv>=1.0.0
requests>=2.31.0






